pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "gunbound"
version = "2.0.0"
description = "GunBound Classic shot calculator with physics simulation and training data"
requires-python = ">=3.10"
dependencies = [
    "Pillow",
]

[project.optional-dependencies]
tools = [
    "pymem",
    "pywin32",
    "PyQt5",
    "numpy",
    "mss",
    "opencv-python",
    "pyautogui",
]

[project.scripts]
gunbound = "gunbound.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
